7.Example/hal/i2c/max30100/Manual/demo_ardunio/extras/recorder/requirements.txt
pyserial
pandas
numpy
scipy
matplotlib
jupyter
